feat: lazy-load older/less-relevant data for orders and jobs
Some views used to eagerly load the complete data from the very beginning where it might rarely be necessary.
In these cases, the period of time is now restricted and can be exntended using a simple "load more" button at the bottom of the view.

For now, it is only implemented for orders and historical jobs.